feat - register emails controller
